telegram style messaging

+++ update-scripts.sh
@@ -1,10 +1,13 @@
 #! /bin/bash
 
-#git fetch origin && git reset --hard origin/master && git clean -f -d
-git fetch origin && git reset --hard origin/dev && git clean -f -d
+if [ "$1" == "master" ]; then
+  echo "Updating to master"
+  git fetch origin && git reset --hard origin/master && git clean -f -d
+fi
+if [ "$1" =="dev" ]; then
+  echo "Updateing to dev"
+  git fetch origin && git reset --hard origin/dev && git clean -f -d
+fi
 
 # Set permissions
 sudo chmod -R 744 *
-
-cd Vsens
-ino clean && ino build && ino upload
